将多个文件中的循环索引变量从 `uint` 改为 `int`
在多个文件中，将循环索引变量从 `uint` 类型更改为 `int` 类型：
- `VKCommandBuffer.cs`：修改 `clearValue.ColorValues.Length` 的循环索引变量，并将 `ColorAttachment` 的赋值从 `i` 改为 `(uint)i`。
- `VKGraphicsContext.PhysicalDevice.cs`：修改 `properties.Length` 的循环索引变量，并将 `graphicsQueueFamilyIndex`、`computeQueueFamilyIndex` 和 `copyQueueFamilyIndex` 的赋值从 `i` 改为 `(uint)i`。
- `VKGraphicsPipeline.cs`：修改 `attachmentCount` 和 `vertexInputBindingCount` 的循环索引变量，并将 `desc.InputLayouts` 的索引从 `(int)i` 改为 `i`，将 `Binding` 的赋值从 `i` 改为 `(uint)i`。
- `VKResourceSet.cs`：修改 `writes.Length` 和 `element.Count` 的循环索引变量。

+++ src/ZenithEngine.Vulkan/VKCommandBuffer.cs
@@ -409,14 +409,14 @@ internal unsafe class VKCommandBuffer : CommandBuffer
 
         if (clearColor)
         {
-            for (uint i = 0; i < clearValue.ColorValues.Length; i++)
+            for (int i = 0; i < clearValue.ColorValues.Length; i++)
             {
                 Vector4D<float> color = clearValue.ColorValues[i];
 
                 ClearAttachment clearAttachment = new()
                 {
                     AspectMask = ImageAspectFlags.ColorBit,
-                    ColorAttachment = i,
+                    ColorAttachment = (uint)i,
                     ClearValue = new()
                     {
                         Color = new()
